feature/GE-630: Broke out to new app, updated tests, added deprecation warnings
